Conflicts:
	ClassicalSharp/GraphicsAPI/Direct3D9Api.cs
	ClassicalSharp/GraphicsAPI/IGraphicsApi.cs
	ClassicalSharp/GraphicsAPI/OpenGLApi.cs
	ClassicalSharp/Program.cs
	OpenTK/GameWindow.cs
	OpenTK/Graphics/IGraphicsContext.cs
	OpenTK/Graphics/OpenGL/GL.cs
	OpenTK/Graphics/OpenGL/GLHelper.cs
	OpenTK/IGameWindow.cs
	OpenTK/INativeWindow.cs
	OpenTK/Interop.cs
	OpenTK/NativeWindow.cs
	OpenTK/Platform/MacOS/CarbonGLNative.cs
	OpenTK/Platform/Windows/WinGLNative.cs
	OpenTK/Platform/X11/X11GLNative.cs

Update the corresponding SDL2-related files in GraphicsAPI as well.

The upstream code removed the ClassicalSharp assembly reference from
the Launcher2 project so that the launcher can be compiled and ran
standalone, but the SDL2 code is in the ClassicalSharp project, which
resulted in a failed build. For now, add back that reference, but it
will be removed in the future (probably the next commit or so) to make
future merges easier.

Also, some graphical glitches were introduced in the DirectX build when
running it under Wine, but I have yet to determine what is responsible
for that.

+++ ClassicalSharp/Map/Formats/MapCw.Exporter.cs
@@ -24,7 +24,7 @@ namespace ClassicalSharp.Map {
 				nbt.Write(NbtTagType.Compound); nbt.Write("ClassicWorld");
 				
 				nbt.Write(NbtTagType.Int8);
-				nbt.Write("FormatVersion"); nbt.WriteInt8(1);
+				nbt.Write("FormatVersion"); nbt.WriteUInt8(1);
 				
 				nbt.Write(NbtTagType.Int8Array);
 				nbt.Write("UUID"); nbt.WriteInt32(16);
@@ -71,11 +71,11 @@ namespace ClassicalSharp.Map {
 			
 			nbt.Write(NbtTagType.Int8);
 			nbt.Write("H");
-			nbt.WriteUInt8((byte)Utils.DegreesToPacked(p.SpawnRotY));
+			nbt.WriteUInt8(Utils.DegreesToPacked(p.SpawnRotY));
 			
 			nbt.Write(NbtTagType.Int8);
 			nbt.Write("P");
-			nbt.WriteUInt8((byte)Utils.DegreesToPacked(p.SpawnHeadX));
+			nbt.WriteUInt8(Utils.DegreesToPacked(p.SpawnHeadX));
 			
 			nbt.Write(NbtTagType.End);
 		}
